[BBC-9033] Item EditableInfo.MDX
